initital commit

+++ tce-scripts/33-tce-install.sh
@@ -18,6 +18,12 @@ if [[ "$EUID" -eq 0 ]]; then
   exit 1
 fi
 
+if [ ! -f ${HOME}/tce/tce-linux-amd64-v${TCE_VERSION}.tar.gz ]; then
+    echo "Expecting Tanzu CLI bundle in ${HOME}/tce/tce-linux-amd64-v${TCE_VERSION}.tar.gz"
+    echo "Exiting ..."
+    exit 1
+fi
+
 rm -rf ${HOME}/.kube-tkg ${HOME}/.kube
 rm -rf ${HOME}/.tanzu ${HOME}/.config/tanzu  ${HOME}.cache/tanzu ${HOME}/.local/share/tanzu-cli ${HOME}/.local/share/tce
 
@@ -25,7 +31,7 @@ rm -rf ${HOME}/.tanzu ${HOME}/.config/tanzu  ${HOME}.cache/tanzu ${HOME}/.local/
 # Install Tanzu Community Edition
 #--------------------------------------------------------------------------------------
 
-echo "Installing Tanzu Community Edition from ${HOME}/tce-linux-amd64-v${TCE_VERSION}"
+echo "Installing Tanzu Community Edition from ${HOME}/tce/tce-linux-amd64-v${TCE_VERSION}.tar.gz"
 cd ${HOME}/tce
 tar xzvf tce-linux-amd64-v${TCE_VERSION}.tar.gz
